requirements.txt
asgiref==3.5.0
Django==4.0.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
git-filter-repo==2.34.0
networkx==2.8.2
pytz==2021.3
sqlparse==0.4.2
